+++ src/tpm2_wrap.c
@@ -2512,7 +2512,7 @@ int wolfTPM2_ReadPCR(WOLFTPM2_DEV* dev, int pcrIndex, int hashAlg, byte* digest,
     PCR_Read_Out pcrReadOut;
     int digestLen = 0;
 
-    if (dev == NULL || pcrIndex < PCR_FIRST || pcrIndex > PCR_LAST)
+    if (dev == NULL || pcrIndex < (int)PCR_FIRST || pcrIndex > (int)PCR_LAST)
         return BAD_FUNC_ARG;
 
     /* set session auth to blank */
